requirements.txt
feedparser==6.0.11
python-dateutil==2.9.0.post0
vaderSentiment==3.3.2
rich>=13.0.0
python-dotenv>=1.0
